requirements.txt
# Core ML frameworks (choose one or both)
tensorflow>=2.13.0
# torch>=2.0.0
# torchvision>=0.15.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Numerical operations
numpy>=1.24.0

# Data handling
pandas>=2.0.0
matplotlib>=3.7.0

# Model utilities
scikit-learn>=1.3.0
